Move output configuration to separate header
